restructure forcefield subpkg
