+++ deeprankcore/domain/forcefield/parsers/vdwparam.py
@@ -1,4 +1,17 @@
-from deeprankcore.models.forcefield.vanderwaals import VanderwaalsParam
+class VanderwaalsParam:
+    def __init__(
+            self,
+            inter_epsilon: float,
+            inter_sigma: float,
+            intra_epsilon: float,
+            intra_sigma: float):
+        self.inter_epsilon = inter_epsilon
+        self.inter_sigma = inter_sigma
+        self.intra_epsilon = intra_epsilon
+        self.intra_sigma = intra_sigma
+    
+    def __str__(self) -> str:
+        return f"{self.inter_epsilon}, {self.inter_sigma}, {self.intra_epsilon}, {self.intra_sigma}"
 
 
 class ParamParser:
